requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
mysql-connector-python==8.2.0
PyJWT==2.8.0
python-dotenv==1.0.0
